# Fix: Improve concepts tree performance