### PR description
This PR provides partial improvements to the tree view performance. However, the primary bottleneck is in the backend, which requires specific endpoints optimized for the tree view, rather than relying on the default concept endpoints. 


### Changes

- **Optimized data retrieval in `ConceptsController#index`:**
The most time-consuming call when expanding a subtree is `@concept = @ontology.explore.single_class({full: true}, params[:id])`
In this PR, it has been updated to `@concept = @ontology.explore.single_class({display: 'prefLabel'}, params[:id])`, as only the prefLabel and id are necessary at this level.
_(Note: This change provides a slight improvement in response time but does not fully resolve the delay.)_

- **Removed redundant sorting:**
Sorting of concept children was duplicated both in the controller and in the component, adding approximately 1 second to load time.