@syphax-bouazzouni 2024-10-30
> @syphax-bouazzouni this branch is already deployed to stage

@Bilelkihal I tested it in Stageportal. The tree performance seems fine, but we still have some latency when we select a concept. The concept takes time to show the details.

[Screencast from 2024-10-30 10-50-52.webm](https://github.com/user-attachments/assets/e339b294-b724-430d-89ac-158d66cc6f6c)