@Bilelkihal 2024-10-30
> Hello @Bilelkihal, can you add a benchmark table with the before and after optimization, with cache enabled, and run it like 5 times, to see if it is better or not? Here is how
> 
> ```ruby
> require 'benchmark'
> 
> iterations = 5
> total_time = 0
> 
> Benchmark.bmbm do |bm|
>   bm.report("tree:") do
>     iterations.times do
>       time = Benchmark.measure do
>         # your code to benchmark
>       end
>       total_time += time.real
>     end
>   end
> end 
> 
> average_time = total_time / iterations
> puts "Average execution time over #{iterations} runs: #{average_time.round(5)} seconds"
> ```
> 
> And finally can you open an issue to describe what optimization need to be done in the API side.


### Benchmark (stage api)
**AGROVOC Tree (cache disabled)**
| Concept         | Before | After |
|-----------------|--------|-------|
| Activities      | 3.05s  | 2.54s |
| Administration  | 2.34s  | 1.28s |
| Analysis        | 1.30s  | 1.70s |
| Application | 2.20s | 1.37s |
| Cleaning | 3.30s | 2.7s |

I don't see a big difference, but in general it's better.